requirements.txt
playwright>=1.45.0
PyYAML>=6.0.1
jsonschema>=4.22.0
python-dotenv>=1.0.1
jinja2>=3.1.4
allure-pytest>=2.13.5
axe-playwright-python>=0.1.6
google-adk>=0.2.0
google-generativeai>=0.7.0
